Cleaning up cli code
