+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/cliutils/CliContext.java
@@ -1,0 +1,4 @@
+package org.hl7.fhir.validation.cliutils;
+
+public class CliContext {
+}
